refactor(gen_reply): extract message input and streaming helpers
