+++ tests/test_gen_reply.py
@@ -12,8 +12,10 @@ from PIL import Image
 import pytest
 from nextcord import File, Embed
 
-from discordbot.cogs.gen_reply import _USAGE_FOOTER_RE, _DISCORD_MESSAGE_LIMIT, ReplyGeneratorCogs
+from discordbot.cogs.gen_reply import ReplyGeneratorCogs
 from discordbot.typings.models import ModelSettings, RouteDecision, RuntimeModelCatalog
+from discordbot.cogs._gen_reply.input import USAGE_FOOTER_RE
+from discordbot.cogs._gen_reply.streaming import DISCORD_MESSAGE_LIMIT, ResponseStreamer
 from discordbot.cogs._gen_reply.exceptions import extract_friendly_error
 
 TEST_LLM_MODEL = "test-llm-model"
@@ -314,10 +316,9 @@ async def test_handle_streaming_allows_missing_output_token_details(
 ) -> None:
     """Regression: LiteLLM may return usage with output_tokens_details=null."""
     del economy_isolated_db
-    cog = ReplyGeneratorCogs.__new__(ReplyGeneratorCogs)
     message = FakeMessage()
 
-    result = await cog._handle_streaming(responses=_stream_events(), message=message)
+    result = await ResponseStreamer(message=message, responses=_stream_events()).stream()
 
     expected = (
         f"hello from stream\n\n-# {TEST_LLM_MODEL} · ⬆ 12 ⬇ 34 · $0.00000000"
@@ -336,7 +337,8 @@ async def test_handle_streaming_continues_long_reply_as_reply_chain(
     message = FakeMessage(content="<@999> explain how long Discord replies are handled")
     body = "x" * 4500
 
-    result = await cog._handle_streaming(
+    result = await ResponseStreamer(
+        message=message,
         responses=_stream_events_from(
             events=[
                 SimpleNamespace(type="response.output_text.delta", delta=body),
@@ -349,8 +351,7 @@ async def test_handle_streaming_continues_long_reply_as_reply_chain(
                 ),
             ]
         ),
-        message=message,
-    )
+    ).stream()
 
     usage_footer = (
         f"\n\n-# {TEST_LLM_MODEL} · ⬆ 1 ⬇ 2 · $0.00000000 · 3 虛擬歡樂豆 (+3 虛擬歡樂豆)"
@@ -358,17 +359,17 @@ async def test_handle_streaming_continues_long_reply_as_reply_chain(
     assert result == f"{body}{usage_footer}"
 
     parent = message.replies[0]
-    assert parent.content == body[:_DISCORD_MESSAGE_LIMIT]
+    assert parent.content == body[:DISCORD_MESSAGE_LIMIT]
 
     first_follow_up = parent.replies[0]
-    assert first_follow_up.content == body[_DISCORD_MESSAGE_LIMIT : _DISCORD_MESSAGE_LIMIT * 2]
+    assert first_follow_up.content == body[DISCORD_MESSAGE_LIMIT : DISCORD_MESSAGE_LIMIT * 2]
 
     second_follow_up = first_follow_up.replies[0]
-    assert second_follow_up.content == f"{body[_DISCORD_MESSAGE_LIMIT * 2 :]}{usage_footer}"
+    assert second_follow_up.content == f"{body[DISCORD_MESSAGE_LIMIT * 2 :]}{usage_footer}"
     assert second_follow_up.replies == []
 
     chain_chunks = [parent.content, first_follow_up.content, second_follow_up.content]
-    assert all(len(chunk) <= _DISCORD_MESSAGE_LIMIT for chunk in chain_chunks)
+    assert all(len(chunk) <= DISCORD_MESSAGE_LIMIT for chunk in chain_chunks)
     assert cog.client.responses.create_models == []
 
 
@@ -377,10 +378,10 @@ async def test_handle_streaming_marks_web_search_from_call_event(
 ) -> None:
     """Verifies native web_search_call events trigger the web reaction."""
     del economy_isolated_db
-    cog = _cog()
     message = FakeMessage()
 
-    await cog._handle_streaming(
+    await ResponseStreamer(
+        message=message,
         responses=_stream_events_from(
             events=[
                 SimpleNamespace(type="response.output_text.delta", delta="answer"),
@@ -394,8 +395,7 @@ async def test_handle_streaming_marks_web_search_from_call_event(
                 ),
             ]
         ),
-        message=message,
-    )
+    ).stream()
 
     assert message.added_reactions == ["🌐"]
 
@@ -410,10 +410,10 @@ async def test_handle_streaming_marks_web_search_from_annotation(
     a search signal.
     """
     del economy_isolated_db
-    cog = _cog()
     message = FakeMessage()
 
-    await cog._handle_streaming(
+    await ResponseStreamer(
+        message=message,
         responses=_stream_events_from(
             events=[
                 SimpleNamespace(type="response.output_text.delta", delta="grounded answer"),
@@ -430,8 +430,7 @@ async def test_handle_streaming_marks_web_search_from_annotation(
                 ),
             ]
         ),
-        message=message,
-    )
+    ).stream()
 
     assert message.added_reactions == ["🌐"]
 
@@ -454,35 +453,35 @@ async def test_gen_reply_message_content_and_attachment_helpers(
     embed.add_field(name="Field", value="Value")
     embed.set_footer(text="Footer")
 
-    assert await cog._get_user_prompt(content="hi <@999>") == "hi"
-    assert "Author" in cog._extract_embed_text(embeds=[embed])
+    assert await cog.input_builder.get_user_prompt(content="hi <@999>") == "hi"
+    assert "Author" in cog.input_builder.extract_embed_text(embeds=[embed])
 
     bot_message = FakeMessage(
         content="answer\n\n-# model · ⬆ 1 ⬇ 2 · $0.0 · +3",
         author=FakeAuthor(bot=True, user_id=999),
     )
-    assert await cog._get_cleaned_content(message=bot_message) == "answer"
-    assert _USAGE_FOOTER_RE.search(string=bot_message.content)
+    assert await cog.input_builder.get_cleaned_content(message=bot_message) == "answer"
+    assert USAGE_FOOTER_RE.search(string=bot_message.content)
 
     embed_message = FakeMessage()
     embed_message.embeds = [embed]
-    assert "Title" in await cog._get_cleaned_content(message=embed_message)
+    assert "Title" in await cog.input_builder.get_cleaned_content(message=embed_message)
 
     system_message = FakeMessage()
     system_message.system_content = "joined"
-    assert await cog._get_cleaned_content(message=system_message) == "joined"
+    assert await cog.input_builder.get_cleaned_content(message=system_message) == "joined"
 
-    assert cog._required_modality(content_type="video/mp4") == "video"
-    assert cog._required_modality(content_type="audio/mpeg") == "audio"
-    assert cog._required_modality(content_type="application/pdf") == "image"
+    assert cog.input_builder.required_modality(content_type="video/mp4") == "video"
+    assert cog.input_builder.required_modality(content_type="audio/mpeg") == "audio"
+    assert cog.input_builder.required_modality(content_type="application/pdf") == "image"
 
-    file_part = await cog._attachment_to_part(
+    file_part = await cog.input_builder.attachment_to_part(
         attachment=FakeAttachment(filename="note.txt", content_type="text/plain", payload=b"abc")
     )
     assert file_part is not None
     assert file_part["file_data"] == "data:text/plain;base64,YWJj"
 
-    image_part = await cog._image_to_part(
+    image_part = await cog.input_builder.image_to_part(
         source=FakeAttachment(
             filename="pixel.png", content_type="image/png", payload=base64.b64decode(_png_b64())
         )
@@ -491,7 +490,7 @@ async def test_gen_reply_message_content_and_attachment_helpers(
     assert image_part["type"] == "input_image"
 
     monkeypatch.setattr(
-        "discordbot.cogs.gen_reply.get_supported_modalities", lambda model_name: {"image"}
+        "discordbot.cogs._gen_reply.input.get_supported_modalities", lambda model_name: {"image"}
     )
     message = FakeMessage()
     message.attachments = [
@@ -508,8 +507,10 @@ async def test_gen_reply_message_content_and_attachment_helpers(
     img_embed = Embed()
     img_embed.set_image(url="https://example.test/image.png")
     message.embeds = [img_embed]
-    monkeypatch.setattr("discordbot.cogs.gen_reply.get_image_data", lambda image_file: _png_b64())
-    parts = await cog._get_attachment_parts(message=message)
+    monkeypatch.setattr(
+        "discordbot.cogs._gen_reply.input.get_image_data", lambda image_file: _png_b64()
+    )
+    parts = await cog.input_builder.get_attachment_parts(message=message)
     assert [part["type"] for part in parts] == ["input_image", "input_image", "input_image"]
 
 
@@ -519,16 +520,17 @@ async def test_gen_reply_processes_history_reference_and_current_messages(
     """Verifies message processing for history, references, and current prompts."""
     cog = _cog()
     monkeypatch.setattr(
-        "discordbot.cogs.gen_reply.get_supported_modalities", lambda model_name: {"text", "image"}
+        "discordbot.cogs._gen_reply.input.get_supported_modalities",
+        lambda model_name: {"text", "image"},
     )
     bot_msg = FakeMessage(content="bot answer", author=FakeAuthor(bot=True, user_id=999))
     user_msg = FakeMessage(content="hello", author=FakeAuthor(user_id=1))
     with_attachment = FakeMessage(content="see file", author=FakeAuthor(user_id=2))
     with_attachment.attachments = [FakeAttachment(filename="note.txt", content_type="text/plain")]
 
-    bot_processed = await cog._process_single_message(message=bot_msg)
-    user_processed = await cog._process_single_message(message=user_msg)
-    attachment_processed = await cog._process_single_message(message=with_attachment)
+    bot_processed = await cog.input_builder.process_single_message(message=bot_msg)
+    user_processed = await cog.input_builder.process_single_message(message=user_msg)
+    attachment_processed = await cog.input_builder.process_single_message(message=with_attachment)
     assert bot_processed["role"] == "assistant"
     assert user_processed["role"] == "user"
     assert attachment_processed["role"] == "user"
@@ -579,13 +581,22 @@ async def test_gen_reply_routes_and_handlers_without_api(monkeypatch: pytest.Mon
     assert isinstance(message.replies[-1].content, str)
     assert message.replies[-1].content.startswith("<@1> caption")
 
-    async def fake_streaming(responses: SimpleNamespace, message: FakeMessage) -> str:
-        """Records the message passed to streaming."""
-        streamed.append(message)
-        return "done"
-
     streamed: list[FakeMessage] = []
-    monkeypatch.setattr(cog, "_handle_streaming", fake_streaming)
+
+    class FakeResponder:
+        """Records the message handed to the streaming responder."""
+
+        def __init__(self, message: FakeMessage, responses: SimpleNamespace) -> None:
+            """Stores the streaming inputs."""
+            self.message = message
+            self.responses = responses
+
+        async def stream(self) -> str:
+            """Records the message and returns placeholder content."""
+            streamed.append(self.message)
+            return "done"
+
+    monkeypatch.setattr("discordbot.cogs.gen_reply.ResponseStreamer", FakeResponder)
     await cog._handle_message_reply(message=message, system_prompt="system", history_limit=2)
     assert cog.client.responses.create_streams[-1] is True
     assert streamed[-1] is message
@@ -611,7 +622,9 @@ async def test_gen_reply_on_message_dispatches_routes(
         """Returns the parametrized route."""
         return route
 
-    async def fake_reaction(message: FakeMessage, emoji: str, previous: str | None = None) -> str:
+    async def fake_reaction(
+        message: FakeMessage, bot_user: object, emoji: str, previous: str | None = None
+    ) -> str:
         """Records reaction state transitions."""
         calls.append(f"reaction:{emoji}")
         return emoji
@@ -631,7 +644,7 @@ async def test_gen_reply_on_message_dispatches_routes(
         calls.append("_handle_message_reply")
 
     monkeypatch.setattr(cog, "_route_message", fake_route)
-    monkeypatch.setattr(cog, "_handle_reaction", fake_reaction)
+    monkeypatch.setattr("discordbot.cogs.gen_reply.update_reaction", fake_reaction)
     monkeypatch.setattr(cog, "_handle_image_reply", fake_image_handler)
     monkeypatch.setattr(cog, "_handle_video_reply", fake_video_handler)
     monkeypatch.setattr(cog, "_handle_message_reply", fake_message_handler)
